add code

+++ src/main/java/pengyi/domain/service/user/driver/DriverService.java
@@ -85,18 +85,49 @@ public class DriverService implements IDriverService {
         Driver driver = this.show(command.getId());
         driver.fainWhenConcurrencyViolation(command.getVersion());
 
-//        Company company = companyService.show(command.getCompany());
+        Company company = companyService.show(command.getCompany());
 
-        driver.setEmail(command.getEmail());
-        driver.setName(command.getName());
-//        driver.setHead(command.getHead());
-//        driver.setCompany(company);
-        driver.setSex(command.getSex());
-//        driver.setLevel(command.getLevel());
-//        driver.setReportCount(command.getReportCount());
+        driver.setCompany(company);
         driver.setDriverType(command.getDriverType());
+        String identityCarPic = command.getIdentityCardPic().replaceAll("img_tmp", "img");
+        String drivingLicencePic = command.getDrivingLicencePic().replaceAll("img_tmp", "img");
+        driver.setIdentityCardPic(identityCarPic);
+        driver.setDrivingLicencePic(drivingLicencePic);
+        Car car = carService.searchByDriver(driver.getId());
+        if (driver.getDriverType() == DriverType.LIMOUSINE) {
+            String travelPic = command.getTravelPic().replaceAll("img_tmp", "img");
+            driver.setTravelPic(travelPic);
+        } else if (driver.getDriverType() == DriverType.GENERATION) {
+            driver.setDrivingLicenceType(command.getDrivingLicenceType());
+            if (null != car) {
+                carService.delete(car);
+            }
+        } else {
+            String businessPic = command.getBusinessPic().replace("img_tmp", "img");
+            String workPic = command.getWorkPic().replaceAll("img_tmp", "img");
+            String travelPic = command.getTravelPic().replaceAll("img_tmp", "img");
+            driver.setBusinessPic(businessPic);
+            driver.setWorkPic(workPic);
+            driver.setTravelPic(travelPic);
+            if (null != car) {
+                car.setCarType(null);
+                carService.update(car);
+            }
+        }
 
         driverRepository.update(driver);
+
+        fileUploadService.move(identityCarPic.substring(identityCarPic.lastIndexOf("/") + 1));
+        fileUploadService.move(drivingLicencePic.substring(drivingLicencePic.lastIndexOf("/") + 1));
+        if (!CoreStringUtils.isEmpty(command.getTravelPic())) {
+            fileUploadService.move(command.getTravelPic().substring(command.getTravelPic().lastIndexOf("/") + 1));
+        }
+        if (!CoreStringUtils.isEmpty(command.getBusinessPic())) {
+            fileUploadService.move(command.getBusinessPic().substring(command.getBusinessPic().lastIndexOf("/") + 1));
+        }
+        if (!CoreStringUtils.isEmpty(command.getWorkPic())) {
+            fileUploadService.move(command.getWorkPic().substring(command.getWorkPic().lastIndexOf("/") + 1));
+        }
 
         return driver;
     }
@@ -113,6 +144,55 @@ public class DriverService implements IDriverService {
     @Override
     public Driver create(Driver driver) {
         driverRepository.save(driver);
+        return driver;
+    }
+
+    @Override
+    public Driver terraceCreate(CreateDriverCommand command) {
+        BaseUser baseUser = baseUserService.searchByUserName(command.getUserName());
+        if (null != baseUser) {
+            throw new ExistException("用户名[" + command.getUserName() + "]已存在");
+        }
+        Company company = companyService.show(command.getCompany());
+
+        String salt = PasswordHelper.getSalt();
+        String password = PasswordHelper.encryptPassword(command.getPassword(), command.getUserName() + salt);
+
+        String identityCarPic = command.getIdentityCardPic().replaceAll("img_tmp", "img");
+        String drivingLicencePic = command.getDrivingLicencePic().replaceAll("img_tmp", "img");
+
+        Role role = roleService.searchByName("driver");
+        Driver driver;
+        if (command.getDriverType() == DriverType.LIMOUSINE) {
+            String travelPic = command.getTravelPic().replaceAll("img_tmp", "img");
+            driver = new Driver(command.getUserName(), password, salt, EnableStatus.ENABLE, new BigDecimal(0), new Date(), role, null, UserType.DRIVER,
+                    command.getName(), null, company, null, 0.0, 0.0, 0.0, 0, false, command.getDriverType(), identityCarPic, drivingLicencePic,
+                    CoreDateUtils.parseDate(command.getStartDriveDate(), "yyyy-MM-dd"), AuthStatus.AUTH_TERRACE, travelPic, null, command.getPhone(), null, null);
+        } else if (command.getDriverType() == DriverType.GENERATION) {
+            driver = new Driver(command.getUserName(), password, salt, EnableStatus.ENABLE, new BigDecimal(0), new Date(), role, null, UserType.DRIVER,
+                    command.getName(), null, company, null, 0.0, 0.0, 0.0, 0, false, command.getDriverType(), identityCarPic, drivingLicencePic,
+                    CoreDateUtils.parseDate(command.getStartDriveDate(), "yyyy-MM-dd"), AuthStatus.AUTH_TERRACE, null, command.getDrivingLicenceType(), command.getPhone(), null, null);
+        } else {
+            String businessPic = command.getBusinessPic().replace("img_tmp", "img");
+            String workPic = command.getWorkPic().replaceAll("img_tmp", "img");
+            String travelPic = command.getTravelPic().replaceAll("img_tmp", "img");
+            driver = new Driver(command.getUserName(), password, salt, EnableStatus.ENABLE, new BigDecimal(0), new Date(), role, null, UserType.DRIVER,
+                    command.getName(), null, company, null, 0.0, 0.0, 0.0, 0, false, command.getDriverType(), identityCarPic, drivingLicencePic,
+                    CoreDateUtils.parseDate(command.getStartDriveDate(), "yyyy-MM-dd"), AuthStatus.AUTH_TERRACE, travelPic, null, command.getPhone(), businessPic, workPic);
+        }
+
+        driverRepository.save(driver);
+        fileUploadService.move(identityCarPic.substring(identityCarPic.lastIndexOf("/") + 1));
+        fileUploadService.move(drivingLicencePic.substring(drivingLicencePic.lastIndexOf("/") + 1));
+        if (!CoreStringUtils.isEmpty(command.getTravelPic())) {
+            fileUploadService.move(command.getTravelPic().substring(command.getTravelPic().lastIndexOf("/") + 1));
+        }
+        if (!CoreStringUtils.isEmpty(command.getBusinessPic())) {
+            fileUploadService.move(command.getBusinessPic().substring(command.getBusinessPic().lastIndexOf("/") + 1));
+        }
+        if (!CoreStringUtils.isEmpty(command.getWorkPic())) {
+            fileUploadService.move(command.getWorkPic().substring(command.getWorkPic().lastIndexOf("/") + 1));
+        }
         return driver;
     }
 
@@ -374,7 +454,9 @@ public class DriverService implements IDriverService {
             driver.setTravelPic(travelPic);
         } else if (driver.getDriverType() == DriverType.GENERATION) {
             driver.setDrivingLicenceType(command.getDrivingLicenceType());
-            carService.delete(car);
+            if (null != car) {
+                carService.delete(car);
+            }
         } else {
             String businessPic = command.getBusinessPic().replace("img_tmp", "img");
             String workPic = command.getWorkPic().replaceAll("img_tmp", "img");
@@ -382,8 +464,10 @@ public class DriverService implements IDriverService {
             driver.setBusinessPic(businessPic);
             driver.setWorkPic(workPic);
             driver.setTravelPic(travelPic);
-            car.setCarType(null);
-            carService.update(car);
+            if (null != car) {
+                car.setCarType(null);
+                carService.update(car);
+            }
         }
 
         driverRepository.update(driver);
